requirements.txt
pandas>=0.24.1
numpy>=1.14.2
scikit-learn>=0.20.3
keras>=2.2.4
PDPbox>=0.2.0
ipython>=6.4.0
scikit-optimize>=0.5.2
python-dateutil>=2.7.3
